% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_button.R
\name{download_button}
\alias{download_button}
\title{Make a groovy download button}
\usage{
download_button(link, text = "Full results")
}
\arguments{
\item{link}{Where the button points}

\item{text}{What the button says}
}
\description{
Must enable font awesome to get the icon; this is set up by default in \code{\link[=knitr_setup]{knitr_setup()}}.
}
