% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_pages.R
\name{gh_pages}
\alias{gh_pages}
\title{Push a flat repository to the GitHub pages branch}
\usage{
gh_pages(dir, files, flatten = FALSE)
}
\arguments{
\item{dir}{The subdirectory containing the web materials to be pushed to \code{gh-pages}}

\item{files}{Optional: a vector of filenames, in the case of a partial push.}

\item{flatten}{Delete git history; this is desirable for the gh-pages branch since its history is irrelevant, but super-dangerous if you specify the wrong directory, so be very careful with this.  Turned off by default.}
}
\description{
Assumes you are in the project home (the one with the \code{.git} directory), and that a subdirectory contains the web materials.
}
\details{
Flat repository means the history is wiped; this is desirable if the html directory is automatically generated.
}
