% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_css.R
\name{include_css}
\alias{include_css}
\title{Creates a css doc}
\usage{
include_css(dir = "_include", quiet = FALSE)
}
\arguments{
\item{dir}{Where to put the child rmd document (default: '_include')}

\item{quiet}{Don't print any output (default: FALSE)}
}
\description{
Creates a css doc
}
\examples{
\dontrun{
include_setup()
}

}
