% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/include_setup.R
\name{include_setup}
\alias{include_setup}
\title{Creates a child rmd with setup and footer code to be used on multiple pages}
\usage{
include_setup(dir = "_include")
}
\arguments{
\item{dir}{Where to put the child rmd document (default: '_include')}
}
\description{
Creates a child rmd with setup and footer code to be used on multiple pages
}
\examples{
\dontrun{
include_setup()
}

}
