% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitr_setup.R
\name{knitr_setup}
\alias{knitr_setup}
\title{Setup script to be called at beginning of rmarkdown files}
\usage{
knitr_setup(fa = TRUE)
}
\arguments{
\item{fa}{Load font awesome? (default: TRUE)}
}
\description{
Loads knitr, kableExtra, font awesome, sets an RNG seed, and sets several
convenient knitr options:
\itemize{
\item Aliases for h (fig.height) and w (fig.width)
\item Turns on fig.align="center", message=FALSE, etc., almost always what one
wants if writing a report (as opposed to an R instructional document)
\item Sets smaller figure margins
}
}
