% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.R
\name{make}
\alias{make}
\title{Build website}
\usage{
make(complete = FALSE, purl = TRUE)
}
\arguments{
\item{complete}{Build entire website from scratch?  By default, only rmd files with newer modification times are built}

\item{purl}{Deposit .R files in _R folder?  Default is true.}
}
\description{
Build website
}
