% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all.R
\name{render_all}
\alias{render_all}
\title{Render all rmd files in the \code{web} directory}
\usage{
render_all(files, quiet = TRUE, ...)
}
\arguments{
\item{files}{Files to render (default: \verb{web/*.rmd} if \code{web} folder exists, otherwise \verb{*.rmd})}

\item{quiet}{As in \code{rmarkdown::render()}; Default is TRUE.}

\item{...}{Further arguments to \code{render_page()}}
}
\description{
Render all rmd files in the \code{web} directory
}
