% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_html.R
\name{render_html}
\alias{render_html}
\title{Render an rmd file built within the html package template}
\usage{
render_html(f, purl = TRUE)
}
\arguments{
\item{f}{File to render}

\item{purl}{Deposit .R files in _R folder?  Default is true.}
}
\description{
Render an rmd file built within the html package template
}
