% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_table.R
\name{render_table}
\alias{render_table}
\title{Render multiple rmd files with input coming from a table}
\usage{
render_table(Table, outDir = "web", layout_dir = "web/_layouts")
}
\arguments{
\item{Table}{Either a \code{data.frame}-like object or a file name (in which case \code{fread()} is used to read the file).  Must contain a column called ID that specifies a unique output filename.}

\item{outDir}{Where to write the output.  Default: web}

\item{layout_dir}{Where are the layouts?  Default: `web/_layouts'}
}
\description{
Render multiple rmd files with input coming from a table
}
