% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_template.R
\name{render_template}
\alias{render_template}
\title{Render a list of variables, including layout, to an rmd file for rendering}
\usage{
render_template(v, out = stdout(), layout_dir = "web/_layouts")
}
\arguments{
\item{v}{A list of variables that the template uses; must contain \code{layout}}

\item{out}{Where to write the output.  Default: stdout}

\item{layout_dir}{Where are the layouts?  Default: `web/_layouts'}
}
\description{
Render a list of variables, including layout, to an rmd file for rendering
}
