% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skeleton.R
\name{skeleton}
\alias{skeleton}
\title{Create web skeleton}
\usage{
skeleton(dir = "web")
}
\arguments{
\item{dir}{Name of directory (default: web)}
}
\description{
Create web skeleton
}
\examples{
\dontrun{
skeleton()
}

}
