% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{demo}
\alias{demo}
\title{Creates a demo folder for illustrating the usage of the package}
\usage{
demo(jekyll = FALSE)
}
\arguments{
\item{jekyll}{Set up in jekyll format (as opposed to \code{docs})? (default: FALSE)}
}
\description{
Creates a demo folder for illustrating the usage of the package
}
\examples{
\dontrun{
demo()
}

}
