% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh_pages.R
\name{gh_pages}
\alias{gh_pages}
\title{Push a flat repository to the GitHub pages branch}
\usage{
gh_pages(dir)
}
\arguments{
\item{dir}{The subdirectory contaiing the web materials to be pushed to \code{gh-pages}}
}
\description{
Assumes you are in the project home (the one with the \code{.git} directory), and that a subdirectory contains the web materials.
}
\details{
Flat repository means the history is wiped; this is desirable if the html directory is automatically generated.
}
