% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_all.R
\name{render_all}
\alias{render_all}
\title{Render all rmd files in the \code{web} directory}
\usage{
render_all(
  files = list.files("web", "*.rmd", full.names = TRUE),
  clean = TRUE,
  purl = TRUE,
  quiet = TRUE
)
}
\arguments{
\item{files}{Files to knit; default: web/*.rmd}

\item{clean}{Remove old files?  Default is TRUE.}

\item{purl}{Deposit .R files in _R folder?  Default is TRUE.}

\item{quiet}{As in \code{rmarkdown::render()}; Default is TRUE.}
}
\description{
Render all rmd files in the \code{web} directory
}
