% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_page.R
\name{render_page}
\alias{render_page}
\title{Render an rmd file built within the html package template}
\usage{
render_page(f, web = FALSE, purl = FALSE, quiet = FALSE)
}
\arguments{
\item{f}{File to render}

\item{web}{For website; assume a workflow in which \code{f="web/index.rmd"} and output to \verb{web/_site} and \verb{web/_R}}

\item{purl}{Deposit .R files in _R folder?  Default is FALSE.}

\item{quiet}{As in \code{rmarkdown::render()}; Default is FALSE.}
}
\description{
Render an rmd file built within the html package template
}
