% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_page.R
\name{render_page}
\alias{render_page}
\title{Render an rmd file built within the html package template}
\usage{
render_page(f, purl = FALSE, quiet = FALSE, force = FALSE)
}
\arguments{
\item{f}{File to render (character)}

\item{purl}{Deposit .r files in _r folder? (default: \code{FALSE})}

\item{quiet}{As in \code{rmarkdown::render()}. (default: \code{FALSE})}

\item{force}{By default, \code{render_page()} requires the folders \code{docs} or \code{web} to exist if multiple \code{.html} files are present. To override this behavior and force rendering, set \code{force=TRUE}.}
}
\description{
Render an rmd file built within the html package template
}
